Test Cleanup, rename namespaces
