+++ BlazorFormSample/Client/Client/GameSystems/GameSystemEditor.razor
@@ -1,7 +1,4 @@
-﻿@using BlazorFormSample.Shared
-@using SpiralAngle.Blazor.Extensions.Forms
-@using BlazorFormSample.Client.SharedComponent
-@using BlazorFormSample.Client.Shared.Editors
+﻿@using BlazorFormSample.Client.SharedComponent
 @using BlazorFormSample.Shared.GameModels
 @using Microsoft.AspNetCore.Authorization
 @attribute [Authorize]
