+++ BlazorFormSample/Client/Client/GameSystems/PropertyCard.razor
@@ -1,5 +1,4 @@
-﻿@using BlazorFormSample.Shared
-@using Humanizer
+﻿@using Humanizer
 @typeparam TEntity
 
 <div class="card">
